requirements.txt
apturl==0.5.2
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
duplicity==0.8.21
elastic-transport==8.1.2
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
fasteners==0.14.1
future==0.18.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
importlib-resources==5.7.1
inflection==0.5.1
install==1.3.5
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.5.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.1.1
protobuf==3.12.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.0
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.6
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
razorpay==1.3.0
reportlab==3.6.8
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
systemd-python==234
twilio==7.9.2
tzdata==2022.1
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.9
usb-creator==0.3.7
wadllib==1.3.6
waitress==2.1.1
whitenoise==6.2.0
xdg==5
xkit==0.0.0
zipp==3.8.0
